add async and repos

+++ app/services/transaction_service.py
@@ -1,49 +1,43 @@
 from sqlalchemy.orm import Session
-from fastapi import HTTPException
-from app.models.transaction import Transaction
+from fastapi import HTTPException, Depends
+
+from app.repositories.transactionrepo import TransactionRepository
 from app.core.redis import redis_client
 from app.core.cache import cache
 
+class TransactionService:
+    def __init__(self, tx_repo: TransactionRepository = Depends()):
+        self.tx_repo = tx_repo
 
-def create_transaction(db: Session, data: dict, user_id: int):
-    tx = Transaction(**data, user_id=user_id)
+    async def create_transaction(self, data: dict, user_id: int):
+        tx = await self.tx_repo.create(data, user_id)
 
-    db.add(tx)
-    db.commit()
-    db.refresh(tx)
-
-    redis_client.flushdb()
-    return tx
-
-
-@cache(ttl=120)
-def get_transactions(db: Session, user_id: int):
-    txs = db.query(Transaction).filter(Transaction.user_id == user_id).all()
-
-    return [
-        {
-            "id": t.id,
-            "title": t.title,
-            "type": t.type,
-            "category": t.category,
-            "amount": t.amount,
-            "created_at": t.created_at
-        }
-        for t in txs
-    ]
+        redis_client.flushdb()
+        return tx
 
 
-def delete_transaction(db: Session, tx_id: int, user_id: int):
-    tx = db.query(Transaction).filter(
-        Transaction.id == tx_id,
-        Transaction.user_id == user_id
-    ).first()
+    @cache(ttl=120)
+    async def get_transactions(self, user_id: int):
+        txs = await self.tx_repo.get_all_transactions(user_id)
 
-    if not tx:
-        raise HTTPException(404, "Transaction not found")
+        return [
+            {
+                "id": t.id,
+                "title": t.title,
+                "type": t.type,
+                "category": t.category,
+                "amount": t.amount,
+                "created_at": t.created_at
+            }
+            for t in txs
+        ]
 
-    db.delete(tx)
-    db.commit()
 
-    redis_client.flushdb()
-    return {"message": "deleted"}
+    async def delete_transaction(self, tx_id: int, user_id: int):
+        deleted = await self.tx_repo.delete(tx_id, user_id)
+
+        if not deleted:
+            raise HTTPException(404, "Transaction not found")
+
+        redis_client.flushdb()
+        return {"message": "deleted"}
